Added mod displayName property [for color tags, etc]
