+++ core/src/io/anuke/mindustry/Vars.java
@@ -118,6 +118,8 @@ public class Vars implements Loadable{
     public static boolean headless;
     /** whether steam is enabled for this game */
     public static boolean steam;
+    /** whether typing into the console is enabled - developers only */
+    public static boolean enableConsole = false;
     /** application data directory, equivalent to {@link io.anuke.arc.Settings#getDataDirectory()} */
     public static FileHandle dataDirectory;
     /** data subdirectory used for screenshots */
