+++ core/src/io/anuke/mindustry/mod/Mods.java
@@ -738,10 +738,14 @@ public class Mods implements Loadable{
 
     /** Plugin metadata information.*/
     public static class ModMeta{
-        public String name, author, description, version, main, minGameVersion;
+        public String name, displayName, author, description, version, main, minGameVersion;
         public Array<String> dependencies = Array.with();
         /** Hidden mods are only server-side or client-side, and do not support adding new content. */
         public boolean hidden;
+
+        public String displayName(){
+            return displayName == null ? name : displayName;
+        }
     }
 
     /** Thrown when an error occurs while loading a mod.*/
