+++ server/src/io/anuke/mindustry/server/ServerLauncher.java
@@ -1,22 +1,58 @@
 package io.anuke.mindustry.server;
 
 
+import io.anuke.arc.*;
 import io.anuke.arc.backends.headless.*;
+import io.anuke.arc.files.*;
+import io.anuke.arc.util.*;
 import io.anuke.mindustry.*;
 import io.anuke.mindustry.core.*;
+import io.anuke.mindustry.mod.*;
+import io.anuke.mindustry.net.Net;
 import io.anuke.mindustry.net.*;
 
-import static io.anuke.mindustry.Vars.platform;
+import static io.anuke.mindustry.Vars.*;
 
-public class ServerLauncher{
+public class ServerLauncher implements ApplicationListener{
+    static String[] args;
 
     public static void main(String[] args){
         try{
+            ServerLauncher.args = args;
             Vars.platform = new Platform(){};
             Vars.net = new Net(platform.getNet());
-            new HeadlessApplication(new MindustryServer(args), null, throwable -> CrashSender.send(throwable, f -> {}));
+            new HeadlessApplication(new ServerLauncher(), null, throwable -> CrashSender.send(throwable, f -> {}));
         }catch(Throwable t){
             CrashSender.send(t, f -> {});
         }
+    }
+
+    @Override
+    public void init(){
+        Core.settings.setDataDirectory(Core.files.local("config"));
+        loadLocales = false;
+        headless = true;
+
+        FileHandle plugins = Core.settings.getDataDirectory().child("plugins");
+        if(plugins.isDirectory() && plugins.list().length > 0 && !plugins.sibling("mods").exists()){
+            Log.warn("[IMPORTANT NOTICE] &lrPlugins have been detected.&ly Automatically moving all contents of the plugin folder into the 'mods' folder. The original folder will not be removed; please do so manually.");
+            plugins.sibling("mods").mkdirs();
+            for(FileHandle file : plugins.list()){
+                file.copyTo(plugins.sibling("mods"));
+            }
+        }
+
+        Vars.loadSettings();
+        Vars.init();
+        content.createBaseContent();
+        mods.loadScripts();
+        content.createModContent();
+        content.init();
+
+        Core.app.addListener(logic = new Logic());
+        Core.app.addListener(netServer = new NetServer());
+        Core.app.addListener(new ServerControl(args));
+
+        mods.each(Mod::init);
     }
 }
